requirements.txt
streamlit
huggingface_hub
pillow
gtts
python-dotenv

